Checking prev-hash is a good idea (de-syncing clocks is not)
Because each node tracks the due_time of the next block itself,
and varies it based on whether the next block is overdue or not,
the clocks drift apart; eventually this leads to two blocks being
minted at the same time, and then half the network de-syncs.

Varying the block time works well as a catch-up mechanism, but the
due_time needs to achieve consensus as well.